// featurehub-webhook-utils/app/models/models/base-rollout-strategy-attribute.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.5
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { RolloutStrategyAttributeConditional } from './rollout-strategy-attribute-conditional';
import { RolloutStrategyFieldType } from './rollout-strategy-field-type';

/**
 * 
 * @export
 * @interface BaseRolloutStrategyAttribute
 */
export interface BaseRolloutStrategyAttribute {
    /**
     * 
     * @type {RolloutStrategyAttributeConditional}
     * @memberof BaseRolloutStrategyAttribute
     */
    conditional?: RolloutStrategyAttributeConditional;
    /**
     * 
     * @type {string}
     * @memberof BaseRolloutStrategyAttribute
     */
    fieldName?: string;
    /**
     * the value(s) associated with this rule
     * @type {Array<any>}
     * @memberof BaseRolloutStrategyAttribute
     */
    values?: Array<any>;
    /**
     * 
     * @type {RolloutStrategyFieldType}
     * @memberof BaseRolloutStrategyAttribute
     */
    type?: RolloutStrategyFieldType;
}